fix(auth): resolve issue with logout confirmation reappearing after logout
