+++ frontend/src/layout/MainLayout.tsx
@@ -1,55 +1,19 @@
 import NavBar from '../components/navbar/NavBar';
-import { NavigateFunction, Outlet, useNavigate } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import Footer from '../components/footer/Footer';
-import { useAppDispatch, useAppSelector } from '../redux/hooks';
+import { useAppSelector } from '../redux/hooks';
 import { RootState } from '../redux/store';
-import { logoutUser } from '../redux/thunk/authThunk';
-import { clearRestaurant } from '../redux/slice/restaurantSlice';
 import { IUser, UserRole } from '../types';
 import { checkRole } from '../utils';
-import { clearAdminDashboard } from '../redux/slice/dashboardSlice';
-import { clearMenus } from '../redux/slice/menusSlice';
-import { clearProfile } from '../redux/slice/profileSlice';
-import { clearCart } from '../redux/slice/cartSlice';
-import { clearOtpTokenTimer } from '../redux/slice/otpTokenSlice';
 
 const MainLayout = () => {
     const currentUser: IUser | null = useAppSelector((store: RootState) => store.authReducer.authData);
-    const naivgate: NavigateFunction = useNavigate();
-    const dispatch = useAppDispatch();
-
-    const isAdmin: boolean = checkRole(UserRole.ADMIN, currentUser?.role);
-    const isRestaurant: boolean = checkRole(UserRole.RESTAURANT, currentUser?.role);
-    const isUser: boolean = checkRole(UserRole.USER, currentUser?.role);
-
-    const handleLogout = async (): Promise<void> => {
-        const response = await dispatch(logoutUser());
-        if (isUser) {
-            dispatch(clearProfile());
-            dispatch(clearMenus());
-            dispatch(clearCart());
-        }
-
-        if (isRestaurant) {
-            dispatch(clearRestaurant());
-            dispatch(clearMenus());
-        }
-
-        if (isAdmin) {
-            dispatch(clearAdminDashboard());
-        }
-        dispatch(clearOtpTokenTimer());
-        // Check if the action was rejected
-        if (response.meta.requestStatus !== 'rejected') {
-            naivgate('/auth');
-        }
-    };
 
     return (
         <div className="flex flex-col min-h-screen w-full">
             {/* Header */}
             <header className="sticky top-0 z-50">
-                <NavBar currentUser={currentUser!} handleLogout={handleLogout} />
+                <NavBar currentUser={currentUser!} />
             </header>
 
             {/* Main body content */}
